+++ air/src/execution_step/execution_context/context.rs
@@ -18,15 +18,21 @@ use super::LastError;
 use super::LastErrorDescriptor;
 use super::Scalars;
 use super::Streams;
+use crate::execution_step::RcSecurityTetraplet;
+use crate::execution_step::ValueAggregate;
 use crate::JValue;
+use crate::UncatchableError;
 
 use air_execution_info_collector::InstructionTracker;
 use air_interpreter_cid::CID;
-use air_interpreter_data::CidStore;
+use air_interpreter_data::CanonCidAggregate;
+use air_interpreter_data::CidInfo;
 use air_interpreter_data::CidTracker;
 use air_interpreter_data::GlobalStreamGens;
 use air_interpreter_data::RestrictedStreamGens;
+use air_interpreter_data::TracePos;
 use air_interpreter_interface::*;
+use polyplets::SecurityTetraplet;
 
 use std::rc::Rc;
 
@@ -55,7 +61,7 @@ pub(crate) struct ExecutionCtx<'i> {
     ///   - at least one of xor subgraphs is completed without an error
     ///   - all of seq subgraphs are completed
     ///   - call executed successfully (executed state is Executed)
-    pub(crate) subgraph_complete: bool,
+    subgraph_completeness: bool,
 
     /// Tracker of all met instructions.
     pub(crate) tracker: InstructionTracker,
@@ -69,8 +75,8 @@ pub(crate) struct ExecutionCtx<'i> {
     /// Tracks all functions that should be called from services.
     pub(crate) call_requests: CallRequests,
 
-    /// Merged CID-to-value dictionaries
-    pub(crate) cid_tracker: CidTracker,
+    /// CID-to-something trackers.
+    pub(crate) cid_state: ExecutionCidState,
 }
 
 impl<'i> ExecutionCtx<'i> {
@@ -88,15 +94,15 @@ impl<'i> ExecutionCtx<'i> {
             current_ingredients.restricted_streams,
         );
 
-        let cid_tracker = CidTracker::from_cid_stores(prev_ingredients.cid_store, current_ingredients.cid_store);
+        let cid_state = ExecutionCidState::from_cid_info(prev_ingredients.cid_info, current_ingredients.cid_info);
 
         Self {
             run_parameters,
-            subgraph_complete: true,
+            subgraph_completeness: true,
             last_call_request_id: prev_ingredients.last_call_request_id,
             call_results,
             streams,
-            cid_tracker,
+            cid_state,
             ..<_>::default()
         }
     }
@@ -109,9 +115,23 @@ impl<'i> ExecutionCtx<'i> {
         self.last_call_request_id += 1;
         self.last_call_request_id
     }
+}
 
-    pub(crate) fn get_value_by_cid(&self, cid: &CID) -> Option<Rc<JValue>> {
-        self.cid_tracker.get(cid)
+impl ExecutionCtx<'_> {
+    pub(crate) fn make_subgraph_incomplete(&mut self) {
+        self.subgraph_completeness = false;
+    }
+
+    pub(crate) fn is_subgraph_complete(&self) -> bool {
+        self.subgraph_completeness
+    }
+
+    pub(crate) fn set_subgraph_completeness(&mut self, subgraph_complete: bool) {
+        self.subgraph_completeness = subgraph_complete;
+    }
+
+    pub(crate) fn flush_subgraph_completeness(&mut self) {
+        self.subgraph_completeness = true;
     }
 }
 
@@ -121,7 +141,71 @@ pub(crate) struct ExecCtxIngredients {
     pub(crate) global_streams: GlobalStreamGens,
     pub(crate) last_call_request_id: u32,
     pub(crate) restricted_streams: RestrictedStreamGens,
-    pub(crate) cid_store: CidStore<JValue>,
+    pub(crate) cid_info: CidInfo,
+}
+
+#[derive(Debug, Default, Clone)]
+pub struct ExecutionCidState {
+    pub(crate) value_tracker: CidTracker<JValue>,
+    pub(crate) tetraplet_tracker: CidTracker<SecurityTetraplet>,
+    pub(crate) canon_tracker: CidTracker<CanonCidAggregate>,
+}
+
+impl ExecutionCidState {
+    fn from_cid_info(prev_cid_info: CidInfo, current_cid_info: CidInfo) -> Self {
+        Self {
+            value_tracker: CidTracker::from_cid_stores(prev_cid_info.value_store, current_cid_info.value_store),
+            tetraplet_tracker: CidTracker::from_cid_stores(
+                prev_cid_info.tetraplet_store,
+                current_cid_info.tetraplet_store,
+            ),
+            canon_tracker: CidTracker::from_cid_stores(prev_cid_info.canon_store, current_cid_info.canon_store),
+        }
+    }
+
+    pub(crate) fn get_value_by_cid(&self, cid: &CID<JValue>) -> Result<Rc<JValue>, UncatchableError> {
+        self.value_tracker
+            .get(cid)
+            .ok_or_else(|| UncatchableError::ValueForCidNotFound("value", cid.clone().into()))
+    }
+
+    pub(crate) fn get_tetraplet_by_cid(
+        &self,
+        cid: &CID<SecurityTetraplet>,
+    ) -> Result<RcSecurityTetraplet, UncatchableError> {
+        self.tetraplet_tracker
+            .get(cid)
+            .ok_or_else(|| UncatchableError::ValueForCidNotFound("tetraplet", cid.clone().into()))
+    }
+
+    pub(crate) fn get_canon_value_by_cid(
+        &self,
+        cid: &CID<CanonCidAggregate>,
+    ) -> Result<ValueAggregate, UncatchableError> {
+        let canon_aggregate = self
+            .canon_tracker
+            .get(cid)
+            .ok_or_else(|| UncatchableError::ValueForCidNotFound("canon aggregate", cid.clone().into()))?;
+        let result = self.get_value_by_cid(&canon_aggregate.value)?;
+        let tetraplet = self.get_tetraplet_by_cid(&canon_aggregate.tetraplet)?;
+
+        let fake_trace_pos = TracePos::default();
+        Ok(ValueAggregate {
+            result,
+            tetraplet,
+            trace_pos: fake_trace_pos,
+        })
+    }
+}
+
+impl From<ExecutionCidState> for CidInfo {
+    fn from(value: ExecutionCidState) -> Self {
+        Self {
+            value_store: value.value_tracker.into(),
+            tetraplet_store: value.tetraplet_tracker.into(),
+            canon_store: value.canon_tracker.into(),
+        }
+    }
 }
 
 use serde::Deserialize;
@@ -160,7 +244,7 @@ impl<'i> Display for ExecutionCtx<'i> {
         writeln!(f, "current peer id: {}", self.run_parameters.current_peer_id)?;
         writeln!(f, "init peer id: {}", self.run_parameters.init_peer_id)?;
         writeln!(f, "timestamp: {}", self.run_parameters.timestamp)?;
-        writeln!(f, "subgraph complete: {}", self.subgraph_complete)?;
+        writeln!(f, "subgraph complete: {}", self.subgraph_completeness)?;
         writeln!(f, "next peer public keys: {:?}", self.next_peer_pks)?;
 
         Ok(())