+++ air/src/preparation_step/mod.rs
@@ -15,7 +15,7 @@
  */
 
 mod errors;
-mod minimal_supported_version;
+mod interpreter_versions;
 mod preparation;
 
 pub use errors::PreparationError;
@@ -23,4 +23,5 @@ pub use errors::PreparationError;
 pub(crate) use preparation::prepare;
 pub(crate) use preparation::PreparationDescriptor;
 
-use minimal_supported_version::min_supported_version;
+use interpreter_versions::interpreter_version;
+use interpreter_versions::min_supported_version;