+++ crates/air-lib/interpreter-data/src/executed_state.rs
@@ -21,6 +21,7 @@ use crate::JValue;
 use crate::TracePos;
 
 use air_interpreter_cid::CID;
+use polyplets::SecurityTetraplet;
 use se_de::par_serializer;
 use se_de::sender_serializer;
 use serde::Deserialize;
@@ -60,8 +61,11 @@ pub enum CallResult {
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub enum ValueRef {
-    Scalar(Rc<CID>),
-    Stream { cid: Rc<CID>, generation: u32 },
+    Scalar(Rc<CID<JValue>>),
+    Stream {
+        cid: Rc<CID<JValue>>,
+        generation: u32,
+    },
 }
 
 /// Let's consider an example of trace that could be produces by the following fold:
@@ -127,10 +131,18 @@ pub struct ApResult {
 }
 
 /// Contains ids of element that were on a stream at the moment of an appropriate canon call.
-#[derive(Debug, Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
 pub struct CanonResult {
-    pub canonicalized_element: JValue,
+    pub tetraplet: Rc<CID<SecurityTetraplet>>,
+    pub values: Vec<Rc<CID<CanonCidAggregate>>>,
+}
+
+/// The type Canon trace CID refers to.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct CanonCidAggregate {
+    pub value: Rc<CID<serde_json::Value>>,
+    pub tetraplet: Rc<CID<SecurityTetraplet>>,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]